perf(memory): combine websocket and tasks backend
