+++ .gitignore
@@ -55,8 +55,6 @@ yarn-error.log*
 
 apps/tasks/tasks.cjs
 apps/tasks/tasks.css
-apps/websocket/wssServer.cjs
-apps/websocket/wssServer.css
 apps/nextjs/.million/
 packages/cli/cli.cjs
