+++ nginx.conf
@@ -7,9 +7,9 @@ http {
         listen 7575;
         listen [::]:7575;
 
-        # Route websockets traffic to port 3001
+        # Route websockets traffic to port 3002
         location /websockets {
-            proxy_pass http://${HOSTNAME}:3001;
+            proxy_pass http://127.0.0.1:3002/websockets;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection "Upgrade";
